refactor(palette): deduplicate owner/repo resolution and formatRelativeDate
decision(gh-utility): getOwnerRepo() を pr.ts に切り出し、issue.ts で再利用
decision(palette): formatRelativeDate を palette 直下の共通ユーティリティに抽出
